Provide a framework for cmd line invocations
Writing a C program using libfluffy for just spontaneous monitoring may
be daunting. For instances when the intention is simply to observe the
events on the filesystem, this framework comes in handy.

fluffyrun spawns an instance that acts on the commands sent through
fluffyctl. fluffyrun serves, fluffyctl controls. fluffyctl provides
options to add/remove watch on the fly. When fluffyrun is invoked, it
waits for instructions from fluffyctl and prints events on the stdout.

Signed-off-by: six-k <raam@tinkershack.in>