[fw] fluffyctl option to reinitiate
Fluffy doesn't miss any action on the fs, nevertheless, a user might
want to reinitate the watch. Is this option required? Can't Fluffy be
stopped and restarted to have the same effect? Well, yes, but the
watches will then have to be set manually. A reinitate call handles it.
And, retains the events choice and other previously set limits as well.

Signed-off-by: Raamsri KP <raam@tinkershack.in>