[fw] Options to support file path globbing
Two new options -W and -I to handle arguments with wildcards.

Users wish to use wildcards for paths. We must get the arguments with
the wildcards and glob it. But for now, let's leave it to the shell to
handle it for us. We will extract the globbed paths from *argv[].

Signed-off-by: Raamsri KP <raam@tinkershack.in>